urls: merge with bounce, test 4
