+++ src/Components/ItemProperties.test.jsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { render, screen, waitFor } from '@testing-library/react'
-import { MockViewer, newMockStringValueElt } from '../utils/Ifc.test'
-import { mockRouted } from '../Routed.test'
+import { MockViewer, newMockStringValueElt } from '../utils/IfcMock.test'
+import { mockRoutes } from '../BaseRoutesMock.test'
 import ItemProperties from './ItemProperties'
 
 
 test('ItemProperties for single element', async () => {
   const testLabel = 'Test node label';
-  const {getByText} = render(mockRouted(
+  const {getByText} = render(mockRoutes(
     <ItemProperties
       viewer={new MockViewer}
       element={newMockStringValueElt(testLabel)} />
@@ -19,7 +19,7 @@ test('ItemProperties for single element', async () => {
 
 test('ItemProperties for single element', async () => {
   const testLabel = 'Test node label';
-  const {getByText} = render(mockRouted(
+  const {getByText} = render(mockRoutes(
     <ItemProperties
       viewer={new MockViewer}
       element={newMockStringValueElt(testLabel)} />
