+++ src/Components/NavTree.test.jsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { render } from '@testing-library/react'
-import { MockViewer, newMockStringValueElt } from '../utils/Ifc.test'
-import { mockRouted } from '../Routed.test'
+import { MockViewer, newMockStringValueElt } from '../utils/IfcMock.test'
+import { mockRoutes } from '../BaseRoutesMock.test'
 import NavTree from './NavTree'
 
 
 test('NavTree for single element', () => {
   const testLabel = 'Test node label';
-  const {getByText} = render(mockRouted(
+  const {getByText} = render(mockRoutes(
     <NavTree
       viewer={new MockViewer}
       element={newMockStringValueElt(testLabel)} />
